Move conventions/extensions fixture to cross version tests
as :internal-integ-testing is not available at runtime for these tests
