+++ subprojects/ide/src/crossVersionTest/groovy/org/gradle/plugins/ide/tooling/r210/ConventionsExtensionsCrossVersionFixture.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.fixtures.compatibility
+package org.gradle.plugins.ide.tooling.r210
 
 import org.gradle.api.JavaVersion
 import org.gradle.util.GradleVersion
